+++ lib/components/task_tile.dart
@@ -27,6 +27,7 @@ class _TaskTileState extends State<TaskTile> {
   @override
   Widget build(BuildContext context) {
     return ListTile(
+      isThreeLine: true,
       onLongPress: () => showDialog<String>(
         context: context,
         builder: (BuildContext context) => AlertDialog(
@@ -52,6 +53,7 @@ class _TaskTileState extends State<TaskTile> {
         style: TextStyle(
             decoration: widget.isChecked ? TextDecoration.lineThrough : null),
       ),
+      subtitle: Text('12.00 pm'),
       trailing: Checkbox(
         activeColor: kAccentColor,
         value: widget.isChecked,
