+++ lib/main.dart
@@ -3,9 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:toddyapp/models/task_data.dart';
 import 'package:toddyapp/screens/login.dart';
-import 'package:toddyapp/screens/registration.dart';
+
 import 'package:toddyapp/screens/tasks_screen.dart';
 import 'package:toddyapp/screens/user_page.dart';
+import 'package:toddyapp/screens/welcome_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -29,7 +30,6 @@ class Toddy extends StatelessWidget {
           '/': (context) => WelcomeScreen(),
           'User': (context) => UserPage(),
           'Login': (context) => Login(),
-          'SingUp': (context) => Registration(),
           'TaskScreen': (context) => TasksScreen(),
         },
       ),
