+++ lib/screens/login.dart
@@ -7,18 +7,6 @@ import 'package:toddyapp/screens/tasks_screen.dart';
 
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-class WelcomeScreen extends StatelessWidget {
-  final currentUser = FirebaseAuth.instance.currentUser;
-
-  @override
-  Widget build(BuildContext context) {
-    if (currentUser != null) {
-      return TasksScreen();
-    }
-    return Login();
-  }
-}
-
 class Login extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
@@ -157,51 +145,8 @@ class LoginState extends State<Login> {
                   SizedBox(
                     height: 20,
                   ),
-                  GoogleSingInButton(),
                 ],
               ),
             )));
-  }
-}
-
-class GoogleSingInButton extends StatelessWidget {
-  const GoogleSingInButton({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      onPressed: () async {
-        try {
-          // ignore: unused_local_variable
-          await GoogleAuth().signInwithGoogle();
-          Navigator.pushNamed(context, 'TaskScreen');
-        } catch (e) {
-          print(e);
-        }
-      },
-      child: Container(
-          padding: EdgeInsets.symmetric(horizontal: 10),
-          width: 250,
-          height: 50,
-          decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(25), color: Colors.white),
-          child: Center(
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                FaIcon(
-                  FontAwesomeIcons.google,
-                  color: kAccentColor,
-                ),
-                Text(
-                  'SIng In with Google',
-                  style: TextStyle(color: kAccentColor),
-                ),
-              ],
-            ),
-          )),
-    );
   }
 }
